Fix circular dependency by merging frames schema back into sequences
- Merged src/lib/db/schema/frames.ts into sequences.ts to eliminate circular import
- Deleted separate frames.ts file
- Updated all imports across codebase to reference sequences.ts
- Reverted routeTree.gen.ts to working version to fix router registration issues

Fixes the 'Cannot access r before initialization' error on Vercel deployment
